Time: 3 ms (47.45%) | Memory: 33.6 MB (37.37%) - LeetSync

+++ 1001-n-repeated-element-in-size-2n-array/n-repeated-element-in-size-2n-array.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int repeatedNTimes(vector<int>& nums) {
         int n=nums.size()/2;
-        vector<long>count(10000,0);
+        vector<int>count(10000,0);
         for(int &i:nums){
             count[i]++;
         }
